Tests/LibWeb/Text/expected/HTML/multi-origin-test-example.txt

Origin count: 3
Origins have different ports: true
areDifferentOrigins works: true
PASS
